dijkstra_slalom_DO_NOT_WORK ver1.2.4

+++ Core/Src/Logic/drive.c
@@ -3,7 +3,7 @@
 #include "logic.h"
 #include "params.h"
 
-void drive_calc_offset(uint8_t dist) {
+void drive_calc_offset(float dist) {
   enable_motor();
   reset_current_position();
